@@ -2,17 +2,15 @@ package main
 
 import (
 	"io"
-	"io/ioutil"
+	"os"
 	"os/exec"
 	"strings"
 
-	"gopkg.in/yaml.v2"
+	"gopkg.in/yaml.v3"
 )
 
-// Values is
 type Values = map[string]interface{}
 
-// Exec is
 func Exec(command string) (string, error) {
 	b, err := exec.Command("bash", "-c", command).Output()
 	if err != nil {
@@ -21,16 +19,14 @@ func Exec(command string) (string, error) {
 	return string(b), err
 }
 
-// ReadFile is
 func ReadFile(path string) (string, error) {
-	b, err := ioutil.ReadFile(path)
+	b, err := os.ReadFile(path)
 	if err != nil {
 		return "", err
 	}
 	return string(b), nil
 }
 
-// FromYaml is
 func FromYaml(str string) (Values, error) {
 	out, err := FromYamlMulti(str)
 	if err != nil {
@@ -42,7 +38,6 @@ func FromYaml(str string) (Values, error) {
 	return out[0], nil
 }
 
-// FromYamlMulti is
 func FromYamlMulti(str string) ([]Values, error) {
 	out := []Values{}
 	decoder := yaml.NewDecoder(strings.NewReader(str))
@@ -61,7 +56,6 @@ OUTER:
 	return out, nil
 }
 
-// ToYaml is
 func ToYaml(v interface{}) (string, error) {
 	data, err := yaml.Marshal(v)
 	if err != nil {
